Rename updater to autoscaler
