+++ src/main/java/com/autotune/Autotune.java
@@ -22,7 +22,7 @@ import com.autotune.analyzer.exceptions.MonitoringAgentNotFoundException;
 import com.autotune.analyzer.exceptions.MonitoringAgentNotSupportedException;
 import com.autotune.analyzer.metadataProfiles.MetadataProfileCollection;
 import com.autotune.analyzer.performanceProfiles.MetricProfileCollection;
-import com.autotune.analyzer.recommendations.updater.RecommendationUpdaterService;
+import com.autotune.analyzer.recommendations.autoscaler.RecommendationUpdaterService;
 import com.autotune.analyzer.utils.AnalyzerConstants;
 import com.autotune.common.datasource.DataSourceCollection;
 import com.autotune.common.datasource.DataSourceInfo;
@@ -36,7 +36,7 @@ import com.autotune.operator.InitializeDeployment;
 import com.autotune.operator.KruizeDeploymentInfo;
 import com.autotune.service.HealthService;
 import com.autotune.service.InitiateListener;
-import com.autotune.analyzer.recommendations.updater.settings.SettingsUpdater;
+import com.autotune.analyzer.recommendations.autoscaler.settings.SettingsUpdater;
 import com.autotune.utils.CloudWatchAppender;
 import com.autotune.utils.KruizeConstants;
 import com.autotune.utils.MetricsConfig;
